feat: create cronjob to sync the data
This cronjob will sync the data from words table to the user_words
table.
It is done that if it runs on localhost will run every minute (testing
purpose) and if it run on prod will at midnight.

Also this commit creates a new app config with the hostname since
locally is taking my computer name.